Timer fix

+++ SubstringSearchAlgorithms/Timer.h
@@ -11,9 +11,9 @@
 class  Timer : public Decorator
 {
 private:
-	unsigned int _time;
+        double _time;
 public:
 	Timer(Substring_Search_Algorithms* w) : Decorator(w), _time{ 0 } {}
 	int Search(const std::string& Line_1, const std::string& Line_2);
-	unsigned int time();
+        double time();
 };
